+++ examples/example_check.py
@@ -42,8 +42,9 @@ def main():
         JS console errors, JS errors, and Python logged errors.
         """
         name = __name__
+        open_browser = False
+
         serverapp_config = {
-            "open_browser": False,
             "base_url": "/foo/",
             "root_dir": osp.abspath(example_dir)
         }