+++ examples/federated/core_package/index.template.js
@@ -1,7 +1,8 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { PageConfig, URLExt } from '@jupyterlab/coreutils';
+import { JupyterLab } from '@jupyterlab/application';
+import { PageConfig } from '@jupyterlab/coreutils';
 
 // Promise.allSettled polyfill, until our supported browsers implement it
 // See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/allSettled
@@ -21,28 +22,7 @@ if (Promise.allSettled === undefined) {
     );
 }
 
-require('./style.js')
-
-function loadScript(url) {
-  return new Promise((resolve, reject) => {
-    const newScript = document.createElement('script');
-    newScript.onerror = reject;
-    newScript.onload = resolve;
-    newScript.async = true;
-    document.head.appendChild(newScript);
-    newScript.src = url;
-  });
-}
-
-async function loadComponent(url, scope) {
-  await loadScript(url);
-
-  // From MIT-licensed https://github.com/module-federation/module-federation-examples/blob/af043acd6be1718ee195b2511adf6011fba4233c/advanced-api/dynamic-remotes/app1/src/App.js#L6-L12
-  await __webpack_init_sharing__('default');
-  const container = window._JUPYTERLAB[scope];
-  // Initialize the container, it may provide shared modules and may need ours
-  await container.init(__webpack_share_scopes__.default);
-}
+import('./style.js');
 
 async function createModule(scope, module) {
   try {
@@ -57,46 +37,32 @@ async function createModule(scope, module) {
 /**
  * The main entry point for the application.
  */
-async function main() {
-  var JupyterLab = require('@jupyterlab/application').JupyterLab;
-  var disabled = [];
-  var deferred = [];
-  var ignorePlugins = [];
-  var register = [];
-
-  // This is all the data needed to load and activate plugins. This should be
-  // gathered by the server and put onto the initial page template.
-  const extension_data = JSON.parse(
-    PageConfig.getOption('federated_extensions')
-  );
+export async function main() {
+  const disabled = [];
+  const deferred = [];
+  const ignorePlugins = [];
+  const pluginsToRegister = [];
 
   const federatedExtensionPromises = [];
   const federatedMimeExtensionPromises = [];
   const federatedStylePromises = [];
 
-  // We first load all federated components so that the shared module
-  // deduplication can run and figure out which shared modules from all
-  // components should be actually used.
-  const extensions = await Promise.allSettled(extension_data.map( async data => {
-    await loadComponent(
-      `${URLExt.join(PageConfig.getOption('fullLabextensionsUrl'), data.name, data.load)}`,
-      data.name
-    );
-    return data;
-  }));
+  // This is all the data needed to load and activate plugins. This should be
+  // gathered by the server and put onto the initial page template.
+  const extensions = JSON.parse(
+    PageConfig.getOption('federated_extensions')
+  );
 
-  extensions.forEach(p => {
-    if (p.status === "rejected") {
-      // There was an error loading the component
-      console.error(p.reason);
-      return;
-    }
+  // The set of federated extension names.
+  const federatedExtensionNames = new Set();
 
-    const data = p.value;
+  extensions.forEach(data => {
     if (data.extension) {
+      federatedExtensionNames.add(data.name);
       federatedExtensionPromises.push(createModule(data.name, data.extension));
     }
     if (data.mimeExtension) {
+      federatedExtensionNames.add(data.name);
       federatedMimeExtensionPromises.push(createModule(data.name, data.mimeExtension));
     }
     if (data.style) {
@@ -134,15 +100,18 @@ async function main() {
     }
   }
 
-  // Handle the registered mime extensions.
+  // Handle the mime extensions.
   const mimeExtensions = [];
-  {{#each jupyterlab_mime_extensions}}
-  try {
-    for (let plugin of activePlugins(require('{{@key}}/{{this}}'))) {
-      mimeExtensions.push(plugin);
+  {{#each mimeExtensions}}
+  if (!federatedExtensionNames.has('{{@key}}')) {
+    try {
+      let ext = require('{{@key}}{{#if this}}/{{this}}{{/if}}');
+      for (let plugin of activePlugins(ext)) {
+        mimeExtensions.push(plugin);
+      }
+    } catch (e) {
+      console.error(e);
     }
-  } catch (e) {
-    console.error(e);
   }
   {{/each}}
 
@@ -158,14 +127,17 @@ async function main() {
     }
   });
 
-  // Handled the registered standard extensions.
-  {{#each jupyterlab_extensions}}
-  try {
-    for (let plugin of activePlugins(require('{{@key}}/{{this}}'))) {
-      register.push(plugin);
+  // Handled the standard extensions.
+  {{#each extensions}}
+  if (!federatedExtensionNames.has('{{@key}}')) {
+    try {
+      let ext = require('{{@key}}{{#if this}}/{{this}}{{/if}}');
+      for (let plugin of activePlugins(ext)) {
+        pluginsToRegister.push(plugin);
+      }
+    } catch (e) {
+      console.error(e);
     }
-  } catch (e) {
-    console.error(e);
   }
   {{/each}}
 
@@ -174,7 +146,7 @@ async function main() {
   federatedExtensions.forEach(p => {
     if (p.status === "fulfilled") {
       for (let plugin of activePlugins(p.value)) {
-        register.push(plugin);
+        pluginsToRegister.push(plugin);
       }
     } else {
       console.error(p.reason);
@@ -186,6 +158,7 @@ async function main() {
      console.error(reason);
     });
 
+  // Create a new JupyterLab object and start it
   const lab = new JupyterLab({
     mimeExtensions,
     disabled: {
@@ -199,11 +172,10 @@ async function main() {
         .map(function (val) { return val.raw; })
     },
   });
-  register.forEach(function(item) { lab.registerPluginModule(item); });
+
+  lab.registerPluginModules(pluginsToRegister);
   lab.start({ ignorePlugins });
   lab.restored.then(() => {
     console.debug('Example started!');
   });
 }
-
-window.addEventListener('load', main);